Start refactoring test functions
